add ROM hack support to the sprite priority override check
The game checks if you're on Airman's stage and sets the background priority bit
on all sprites (except health bars) if so. This value is hardcoded into the
main graphics routine (presumably it would've been an assembler macro back in
'88, but whatever). By reading this address, I'm essentially deferring to the
ROM's hardcoding rather than my own!